requirements.txt
# Core framework:
# - Django powers the web application.
Django>=5.2,<6.0

# Django tooling:
# - Matches the installed app listed in `INSTALLED_APPS`.
django-extensions>=4.1,<5.0

# REST framework:
# - Used for the subscription-based article API.
djangorestframework>=3.15,<4.0

# Database driver:
# - Switch to this when you migrate from SQLite to MariaDB.
mariadb>=1.1

# HTTP client:
# - Useful later for calling the X/Twitter API from the approval workflow.
requests>=2.32,<3.0
